Move constants file out of “types”
