+++ server/data/availableComponent.ts
