+++ server/routes/componentRoutes.ts
@@ -4,7 +4,7 @@ import { expressjwt, GetVerificationKey } from 'express-jwt'
 import jwt from 'jsonwebtoken'
 import { Services } from '../services'
 import config from '../config'
-import { isComponent } from '../@types/AvailableComponent'
+import { isComponent } from '../data/availableComponent'
 import { TokenData } from '../@types/Users'
 import type { AvailableComponent, Component, Components } from '../interfaces/externalContract'
 import ComponentsController from '../controllers/componentsController'
