+++ server/routes/developRoutes.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express'
 import { Services } from '../services'
 import authorisationMiddleware from '../middleware/authorisationMiddleware'
-import { AVAILABLE_COMPONENTS } from '../@types/AvailableComponent'
 import auth from '../authentication/auth'
+import { AVAILABLE_COMPONENTS } from '../data/availableComponent'
 import tokenVerifier from '../data/tokenVerification'
 import type { AvailableComponent } from '../interfaces/externalContract'
 import ComponentsController from '../controllers/componentsController'
